+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.2.1
 
 * feat: 依赖库升级
+* feat: WrapperCachedNetworkImage新增defaultPlaceholder属性
 
 ## 1.2.0